ntru.py: updated doctests. Changed *_parameters.homogeneous to _homogeneous
Doctests were changed in accordance with the updated homogeneity checks. Previously,
the +1 on line 261 of lwe_primal.py was removed, with the assumption that a homogeneous
instance would not need the additional dimension. This was a mistake, and improperly
gimping the dimension optimization functionality for homogeneous instances. This
commit rectifies this mistake by updating the doctests to be in line with the more
efficient implementation.