Use is_sparse, specialize to SparseTernary inside
Now, use the general NoiseDistribution.is_sparse method to do a different cost estimation for sparse secrets. But, when it is sparse ternary, use the splitting methods from that noise distribution to get the precise costs.

In the case of lwe_dual, there was only code that worked for SparseTernary, so crash if other objects are put in.
In the case of lwe_guess, use an estimate, i.e. the given for all values in the bounded range the probability to be non-zero is assumed to be the same.

See: https://github.com/malb/lattice-estimator/pull/127#discussion_r1768979784